requirements.txt
requests
pandas
psycopg2-binary
python-dotenv
apache-airflow-providers-postgres
pytest
